moved DotMinecraftDirectoryService
