+++ src/main/java/amidst/mojangapi/file/json/launcherprofiles/LauncherProfileJson.java
@@ -9,9 +9,9 @@ import amidst.documentation.GsonConstructor;
 import amidst.documentation.Immutable;
 import amidst.documentation.NotNull;
 import amidst.mojangapi.MojangApi;
+import amidst.mojangapi.file.DotMinecraftDirectoryService;
 import amidst.mojangapi.file.directory.ProfileDirectory;
 import amidst.mojangapi.file.directory.VersionDirectory;
-import amidst.mojangapi.file.json.DotMinecraftDirectoryService;
 import amidst.mojangapi.file.json.ReleaseType;
 
 @Immutable
