+++ src/test/java/amidst/devtools/MinecraftVersionCompatibilityChecker.java
@@ -11,8 +11,8 @@ import amidst.clazz.Classes;
 import amidst.clazz.real.JarFileParsingException;
 import amidst.clazz.symbolic.declaration.SymbolicClassDeclaration;
 import amidst.clazz.translator.ClassTranslator;
+import amidst.mojangapi.file.DotMinecraftDirectoryService;
 import amidst.mojangapi.file.FilenameService;
-import amidst.mojangapi.file.json.DotMinecraftDirectoryService;
 import amidst.mojangapi.file.json.versionlist.VersionListEntryJson;
 import amidst.mojangapi.file.json.versionlist.VersionListJson;
 import amidst.mojangapi.minecraftinterface.local.DefaultClassTranslator;
